+++ Model/Adapter/StancerAdapter.php
@@ -134,6 +134,8 @@ class StancerAdapter
      * Submit Transaction for authorize
      *
      * @since 1.0.0
+     * @since 1.0.1 Mobile phone is temporary removed
+     * @since 1.0.2 Mobile phone is back
      *
      * @param array $attributes
      * @return Payment|null
@@ -150,6 +152,7 @@ class StancerAdapter
 
         $customer = new Customer();
         $customer->setEmail($attributes[CustomerDataBuilder::CUSTOMER][CustomerDataBuilder::EMAIL]);
+        $customer->setMobile($attributes[CustomerDataBuilder::CUSTOMER][CustomerDataBuilder::PHONE]);
         $customer->setName(implode(' ', [
             $attributes[CustomerDataBuilder::CUSTOMER][CustomerDataBuilder::FIRST_NAME],
             $attributes[CustomerDataBuilder::CUSTOMER][CustomerDataBuilder::LAST_NAME],