+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-09-24
+
+### Changed
+- Use International phone number as customer mobile (CMS-102)
+
+### Fixed
+- Phone number are formatted following E164 standard ([GitHub#1](https://github.com/wearestancer/cms-magento/pull/1)), thanks [@nola-adrien](https://github.com/nola-adrien)
+
+
 ## [1.0.1] - 2024-08-13
 
 ### Removed